load assets in setup
