+++ data/mobs.ron
@@ -1,48 +1,46 @@
-(
-	mobs: {
-		Enemy(Drone): (
-			mob_type: Enemy(Drone),
-			acceleration: (0.0, 0.02),
-			deceleration: (0.09, 0.01),
-			speed: (0.0, 1.0),
-			collider_dimensions: (14.0, 14.0),
-			sprite_dimensions: (18.0, 18.0),
-			texture_path: "texture/drone_spritesheet.png",
-			texture_atlas_cols: 3,
-			texture_atlas_rows: 1,
-		),
-		Enemy(Pawn): (
-			mob_type: Enemy(Pawn),
-			acceleration: (0.0, 0.02),
-			deceleration: (0.09, 0.01),
-			speed: (0.0, 0.85),
-			collider_dimensions: (14.0, 14.0),
-			sprite_dimensions: (18.0, 18.0),
-			texture_path: "texture/pawn_spritesheet.png",
-			texture_atlas_cols: 3,
-			texture_atlas_rows: 1,
-		),
-		Ally(Hauler): (
-			mob_type: Ally(Hauler),
-			acceleration: (0.0, 0.02),
-			deceleration: (0.09, 0.01),
-			speed: (0.0, 0.7),
-			collider_dimensions: (18.0, 32.0),
-			sprite_dimensions: (22.0, 36.0),
-			texture_path: "texture/hauler_spritesheet.png",
-			texture_atlas_cols: 3,
-			texture_atlas_rows: 1,
-		),
-		Enemy(MissileLauncher): (
-			mob_type: Enemy(MissileLauncher),
-			acceleration: (0.0, 0.02),
-			deceleration: (0.09, 0.01),
-			speed: (0.0, 0.75),
-			collider_dimensions: (17.0, 22.0),
-			sprite_dimensions: (21.0, 26.0),
-			texture_path: "texture/missile_launcher_spritesheet.png",
-			texture_atlas_cols: 3,
-			texture_atlas_rows: 1,
-		),
-	}
-)
+{
+	Enemy(Drone): (
+		mob_type: Enemy(Drone),
+		acceleration: (0.0, 0.02),
+		deceleration: (0.09, 0.01),
+		speed: (0.0, 1.0),
+		collider_dimensions: (14.0, 14.0),
+		sprite_dimensions: (18.0, 18.0),
+		texture_path: "texture/drone_spritesheet.png",
+		texture_atlas_cols: 3,
+		texture_atlas_rows: 1,
+	),
+	Enemy(Pawn): (
+		mob_type: Enemy(Pawn),
+		acceleration: (0.0, 0.02),
+		deceleration: (0.09, 0.01),
+		speed: (0.0, 0.85),
+		collider_dimensions: (14.0, 14.0),
+		sprite_dimensions: (18.0, 18.0),
+		texture_path: "texture/pawn_spritesheet.png",
+		texture_atlas_cols: 3,
+		texture_atlas_rows: 1,
+	),
+	Ally(Hauler): (
+		mob_type: Ally(Hauler),
+		acceleration: (0.0, 0.02),
+		deceleration: (0.09, 0.01),
+		speed: (0.0, 0.7),
+		collider_dimensions: (18.0, 32.0),
+		sprite_dimensions: (22.0, 36.0),
+		texture_path: "texture/hauler_spritesheet.png",
+		texture_atlas_cols: 3,
+		texture_atlas_rows: 1,
+	),
+	Enemy(MissileLauncher): (
+		mob_type: Enemy(MissileLauncher),
+		acceleration: (0.0, 0.02),
+		deceleration: (0.09, 0.01),
+		speed: (0.0, 0.75),
+		collider_dimensions: (17.0, 22.0),
+		sprite_dimensions: (21.0, 26.0),
+		texture_path: "texture/missile_launcher_spritesheet.png",
+		texture_atlas_cols: 3,
+		texture_atlas_rows: 1,
+	),
+}
